+++ Tests/Editor/TestData/AVAssemb/AsmdefWithTypeAdd/AFolder/Test.cs
@@ -1,5 +1,0 @@
-#if UNITY_2018_1_OR_NEWER
-public class Foo2
-{
-}
-#endif