+++ Tests/Editor/TestData/AVAssemb/InternalOnlyAsmdefAdd/AFolder/Test.cs
@@ -1,5 +1,0 @@
-#if UNITY_2018_1_OR_NEWER
-internal class InternalFoo
-{
-}
-#endif