+++ Tests/Editor/TestData/AVAssemb/PropAdd/Test.cs
@@ -1,7 +1,0 @@
-#if UNITY_2018_1_OR_NEWER
-public class Foo
-{
-    public int Bar { get; set; }
-    public int Baz { get; set; }
-}
-#endif