+++ Tests/Editor/TestData/AVAssemb/WithObsoleteUpdater/Test.cs
@@ -1,7 +1,0 @@
-#if UNITY_2018_1_OR_NEWER
-public class Foo
-{
-    [System.Obsolete("reason (UnityUpgradable) -> Foo", false)]
-    public int Bar { get; set; }
-}
-#endif