+++ Tests/Editor/Utilities/ReleaseType.cs
@@ -1,9 +1,0 @@
-namespace UnityEditor.PackageManager.ValidationSuite.Tests
-{
-    public enum ReleaseType
-    {
-        Major,
-        Minor,
-        Patch
-    }
-}